backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
cloudinary>=1.36
dj-database-url>=2.1
gunicorn>=21.2
psycopg2-binary>=2.9
python-dotenv>=1.0
whitenoise>=6.6
